0.3.2: Better exception handling when sending handoff handles
In very rare cases, a socket will drop between the time we check it for
disconnect and when Node's child_process code goes to prepare it for
flight.  Catch that case (which bubbles up as a TypeError) and re-try
the send without the handle, letting the receiver decide what to do
with the missing handle.